backend/requirements.txt
# EduReach Backend Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1

# Alternative: FastAPI (uncomment if using FastAPI instead)
# fastapi==0.104.1
# uvicorn==0.24.0
# python-multipart==0.0.6

# Database
pymongo==4.6.0
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL

# AI & ML
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
nltk==3.8.1

# Speech Recognition
faster-whisper==0.10.0
# torch==2.1.1  # Required for faster-whisper

# Translation
deep-translator==1.11.4
gtts==2.4.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Image Processing (for proctoring)
opencv-python==4.8.1.78
Pillow==10.1.0

# File Handling
python-magic==0.4.27
werkzeug==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# Task Queue (optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0

# Monitoring
flask-healthz==1.0.1

# Security
bcrypt==4.1.1
cryptography==41.0.7
